requirements.txt
flask
Pillow
torch
torchvision